Доделал 7 задание
