+++ js/form.js
@@ -1,0 +1,26 @@
+const announcementForm = document.querySelector('.ad-form');
+const announcementFormFieldset = announcementForm.querySelectorAll('fieldset');
+const filterForm = document.querySelector('.map__filters');
+const filterFormFieldset = filterForm.querySelectorAll('fieldset');
+
+const DISABL_CSS_FORM = 'ad-form--disabled';
+
+const disabledForm = (form, cssClass) => {
+  form.classList.add(cssClass);
+};
+
+const disabledFieldset = (allFildset) => {
+  allFildset.forEach((elem) => {
+    elem.disabled = true;
+  });
+};
+
+disabledForm(announcementForm, DISABL_CSS_FORM);
+disabledFieldset(announcementFormFieldset);
+
+disabledForm(filterForm, DISABL_CSS_FORM);
+disabledFieldset(filterFormFieldset);
+
+// Форма заполнения информации об объявлении .ad-form содержит класс ad-form--disabled;
+// Все интерактивные элементы формы .ad-form должны быть заблокированы с помощью атрибута disabled, добавленного на них или на их родительские блоки fieldset;
+// Форма с фильтрами .map__filters заблокирована так же, как и форма .ad-form — на форму добавлен специальный класс, а на её интерактивные элементы атрибуты disabled;
