+++ js/main.js
@@ -1,7 +1,10 @@
+import {getGeneratedAd} from './data.js';
 import {getArrayOfDeclarations} from './util.js';
-import './element-markup.js';
+import './form.js';
 
 const NUMBER_OF_ADS = 10;
 
-getArrayOfDeclarations(NUMBER_OF_ADS);
+const ads =getArrayOfDeclarations(NUMBER_OF_ADS);
 
+const body = document.querySelector('body');
+body.appendChild(getGeneratedAd(ads[0]));
